refactor: input,label 컴포넌트를 카드 등록 페이지 내에서만 공유
