+++ src/feature/CardRegister/components/CardRegisterForm/shared/Label/Label.stories.tsx
